move api definitions to own folder
