+++ crates/rotala/src/exchange/mod.rs
@@ -1,2 +1,1 @@
-pub (crate) mod traits;
 pub mod uist;
