+++ crates/rotala/src/lib.rs
@@ -1,3 +1,4 @@
+mod api;
 mod input;
 mod exchange;
 mod orderbook;
